SQL to create tables, some type changes for metrics and policy settings

+++ app.py
@@ -23,8 +23,8 @@ class Metrics(BaseModel):
     inflationRate: float
     consumerConfidence: float
     investmentLevel: float
-    tradeBalance: int
-    governmentDebt: int
+    tradeBalance: float
+    governmentDebt: float
     unemploymentRate: float
     populationGrowth: float
